refactor: fix mobile view
